Use Helm 2.14.3 instead of 2.11.0
When running the cluster build script via the tools image, we
noticed intermittent failures with prometheus and cert-manager.

The working theory is that custom resource definitions (CRDs) were
sometimes not ready, when they were referenced, perhaps because
the build process was running a little slower, via a docker
container.

Helm 2.14 is reportedly better at handling CRDs than previous
versions, and two test builds in a row were successful, when using
Helm 2.14.

For this reason, and because it's generally good practice to keep
our tools reasonably up to date, this commit upgrades helm to the
latest version.

Note: users who try to use this version of helm to talk to older
tiller installations may be prompted to upgrade their tiller.