Add and switch to a non-root user
The kubernetes cluster will not allow containers
whose user is running as root.

This commit changes the tools image so that it
runs as a non-root user.

Note: to use this image to build a cluster, AWS &
gpg credentials need to be mounted into
/home/toolsuser, not /root

Also, the infrastructure repository code needs to
be mounted into a directory below /home/toolsuser,
not directly into /home/toolsuser. Otherwise, you
end up nuking /home/toolsuser/.terraform.d and the
auth0 plugin, which breaks the cluster build script.