Add the octokit gem
This gem is used by our github actions. Every action we have
installs the octokit gem during its setup.

This commit adds the octokit gem to the tools image, so that this
step will no longer be necessary.

After this commit is merged, the tool-image pipeline should be
updated to tag the image as version 1.5