Add the 'circleci' shell commands
We have our main tools image on docker hub, but a couple of teams
are using a slightly different version, pushed to ECR by a
concourse pipeline.

The 'circleci' version includes a couple of shell scripts to aid
in authenticating to the cluster and pushing docker images. One or
two teams are using this version (based on a github search for the
ECR docker image reference).

This change adds these shell scripts to our 'official' tools image
so that anyone using it can use our version from docker hub,
instead of the ECR version (which will subsequently be removed,
along with the pipeline job that creates it).