Change add "latest" tag to tools image
A later change will push a second docker image,
based on this one, but with the gems for the env.
pipeline ruby scripts pre-installed.

This second docker image will be based on this
image, and adding the "latest" tag to the 'base'
image means the second image should always be
based on the latest version of the tools-image.