Build and push a second image to docker hub
This change adds a second docker image, based on
the first but with the gems required for the
environments pipeline concourse jobs pre-installed

Using this image in the environments pipeline
should save 1 to 2 minutes of execution time for
every plan, apply-changes and apply pipeline job.

This image is built as a subsequent "step", not a
separate job, to ensure that the main tools-image
is built, pushed and tagged *before* we try to 
build another image based on it.

This should mean the latest changes to the tools
image. are incorporated into the pipeline-tools
image.

NB: If any changes are made to the gems listed in
the environments repository Gemfile, a new release 
of this project should be created, and the
pipeline should be updated to use that version of 
the pipeline tools image.

Pipeline jobs will still work, even after such a
change, because (at the time of writing), the
pipeline job still has a `bundle install` step,
which will install any missing gems (running
`bundle install` a second time, with no changes,
is a very quick operation). But, actually
installing more gems within the pipeline job
itself defeats the purpose of this change.