Add multiple Terraform versions to make upgrading easier
We have a huge monolith of Terraform state that's difficult to upgrade
in one go. By having multiple binaries in this tools image allows us to
use conditional formatting to phase an upgrade over a number of days.

For example, conceptually it would look like:
```
if versions.tf contains 0.13
	execute terraform0.13 apply
end
```