Install Terraform 0.13, 0.14 and revert the Pingdom change
This commit connects to
https://github.com/ministryofjustice/cloud-platform/issues/2723 and
relates to upgrading our Terraform stack to 0.13 and subsequently 0.14.

By having all three Terraform binaries in the image, we can perform a
slow upgrade of our Terraform stack, which some state on 0.12 and some
on 0.13 or 0.14. This isn't a long lived change. When the upgrade is
finished the 0.13 and 0.12 binaries will be removed.